Create invoice exporter mixin
